Double precision Orbit and time
The periapsisRadius, eccentricity, and timeOfPeriapsisPassage fields in the Orbit class were updated to use double precision floats instead of single precision. This led to several other changes throughout the codebase.
Other places in the codebase where variables represented periods of time were also updated to use type double. I did a pretty comprehensive search for them all but still might've missed some, but hopefully not.
All the unit tests (apart from the Lambert Solver, which is the entire point of this branch) are still passing.