Uniform behaviour across plotting scripts enforced
It is now uniformly the responsibility of the plotting scripts (OrbitPlot and TransitionStepPlot) to do the checking for whether a plot point is greater than the maximum plotting distance, and so shouldn't be plotted.
It is also now uniformly the responsibility of an external caller to do any necessary rescaling of a plot point. The plotting scripts will just plot what they're given, and no external caller will get a surprise rescaling.