Preparing for OrbitalStep = null to represent invalid transfer orbits
If a ManeuverTransitionStep doesn't have a valid adjacent transfer orbit, then its position when plotting should be Vector3.zero.
I must store the state of the validity of a transfer OrbitalStep, and I believe the best way to do this is to have the "orbit" member in an OrbitalStep be set to "null" when a valid transfer orbit cannot be found, and then to have adjacent ManeuverTransitionSteps react to that appropriately.
The OrbitalStep class is unprepared for the 'orbit = null' case, and these changes make it more ready. The changes also meant a more logical distribution of responsibilities when it comes to plotting an OrbitalStep, with the heavy lifting moving over to the OrbitPlot class.