Changed NamedColour to struct - to make it a value type
A bug when loading the colours of OrbitalSteps was caused because NamedColour being a reference type made the OrbitalStep.stepColour variable refer to the TimelineStepSaveData.stepColour variable that was being repeatedly overwritten by the while-loop in MissionTimeline.SetMissionTimelineFromJsonRepresentation().

NamedColour would be more appropriate as a value type, so that's is how I fixed the bug. This required changing a call to a generic method in OrbitalStep that has the 'is class' constraint applied. The fix was a simple case of appropriate boxing/unboxing.